updating existing encrypted passwords is now possible

+++ encrypt.py
@@ -7,7 +7,7 @@ import hmac
 import base64
 import json
 
-def encrypt(Master_pass,Password,Service,Mail):
+def encrypt(Master_pass:bytes,Password: bytes,Service:str,Mail:str,count_given:int,Update_Existing:bool):
     Nonce = os.urandom(12)# will be saved with the password, 2 different Nonces will be generated
 
     key = open("examplekey.txt").read() #Will be saved on a hard drive
@@ -26,21 +26,49 @@ def encrypt(Master_pass,Password,Service,Mail):
     final_key =  hmac.new(key=key, msg=Derived_Master, digestmod=hashlib.sha256).digest() # will not be saved
     aesgcm = AESGCM(final_key) # will not be saved
     encryptedpassword = aesgcm.encrypt(Nonce,Password,aad) # will be saved
+    count = 1
 
 
 
-    Datatosave = {
-        "Service":Service,
-        "Mail": Mail,
-        "Nonce":base64.b64encode(Nonce).decode(),
-        "salt":base64.b64encode(salt).decode(),
-        "iterations":KDFIterations,
-        "ciphertext":base64.b64encode(encryptedpassword).decode(),
-        "aad": base64.b64encode(aad).decode()
-                }
+    
     with open("exampledata.json") as f:
         Database = json.load(f)
-
-    Database["Entries"].append(Datatosave)
-    with open("exampledata.json", "w") as f:
-        json.dump(Database, f, indent=2)
+    
+    Datatosave = {
+                    "Service":Service,
+                    "Mail": Mail,
+                    "Nonce":base64.b64encode(Nonce).decode(),
+                    "salt":base64.b64encode(salt).decode(),
+                    "iterations":KDFIterations,
+                    "ciphertext":base64.b64encode(encryptedpassword).decode(),
+                    "aad": base64.b64encode(aad).decode(),
+                    "count":0
+                            }
+    for entry in Database["Entries"]:
+        if entry["Service"] == Service and entry["Mail"]== Mail :
+            if Update_Existing == False:
+                count+=1
+                
+            elif entry["count"] == count_given:
+                index = Database["Entries"].index(entry)
+                entry["ciphertext"] = base64.b64encode(encryptedpassword).decode()
+                entry["salt"] = base64.b64encode(salt).decode() 
+                entry["Nonce"] = base64.b64encode(Nonce).decode()
+                entry["aad"] = base64.b64encode(aad).decode()
+                print(Master_pass)
+                
+                
+                
+                break
+                
+    
+                
+        
+    if not Update_Existing:
+        Datatosave["count"] = count
+        Database["Entries"].append(Datatosave)
+        with open("exampledata.json", "w") as f:
+            json.dump(Database, f, indent=2)
+    else:
+        with open("exampledata.json", "w") as f:
+            json.dump(Database, f, indent=2)
